update README and images
rename cache to be more accurate

create a new utils folder
